replace old withDispatch with new
